+++ src/main/java/zupzup/back_end/reservation/dto/OrderDto.java
@@ -3,7 +3,7 @@ package zupzup.back_end.reservation.dto;
 import lombok.Getter;
 import lombok.Setter;
 import org.modelmapper.ModelMapper;
-import zupzup.back_end.store.domain.Order;
+import zupzup.back_end.reservation.domain.Order;
 import zupzup.back_end.store.domain.type.OrderStatus;
 import zupzup.back_end.reservation.dto.request.OrderRequestDto;
