+++ src/main/java/zupzup/back_end/reservation/repository/OrderRepository.java
@@ -2,7 +2,7 @@ package zupzup.back_end.reservation.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import zupzup.back_end.store.domain.Order;
+import zupzup.back_end.reservation.domain.Order;
 
 @Repository
 public interface OrderRepository extends JpaRepository<Order, Long> {
